+++ src/main/java/com/danrus/rpf/logging/ModelTestsResultCollector.java
@@ -1,6 +1,6 @@
 package com.danrus.rpf.logging;
 
-import net.minecraft.resources.Identifier;
+import net.minecraft.resources.ResourceLocation;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
@@ -9,27 +9,27 @@ import java.util.List;
 public class ModelTestsResultCollector {
     private final List<TestResultUnit> units = new LinkedList<>();
 
-    public void touchModelNotFound(Identifier location) {
+    public void touchModelNotFound(ResourceLocation location) {
         touch("Not Found", "Unknown", location, TestResultType.ERROR);
     }
 
-    public void touchAllow(String itemModelType, String packName, Identifier location) {
+    public void touchAllow(String itemModelType, String packName, ResourceLocation location) {
         touch(itemModelType, packName, location, TestResultType.ALLOW_UPDATE);
     }
 
-    public void touchDelegate(String itemModelType, String packName, Identifier location) {
+    public void touchDelegate(String itemModelType, String packName, ResourceLocation location) {
         touch(itemModelType, packName, location, TestResultType.DELEGATE);
     }
 
-    public void touchNext(String itemModelType, String packName, Identifier location) {
+    public void touchNext(String itemModelType, String packName, ResourceLocation location) {
         touch(itemModelType, packName, location, TestResultType.NEXT_TEST);
     }
 
-    private void touch(String itemModelType, String packName, Identifier location, TestResultType resultType) {
+    private void touch(String itemModelType, String packName, ResourceLocation location, TestResultType resultType) {
         units.add(new TestResultUnit(itemModelType, packName, location, resultType));
     }
 
-    public Identifier getModelLocation() { // FIXME: hack, i made this under my beer
+    public ResourceLocation getModelLocation() { // FIXME: hack, i made this under my beer
         return units.getFirst().itemModel();
     }
 
@@ -41,7 +41,7 @@ public class ModelTestsResultCollector {
         return strings;
     }
 
-    private record TestResultUnit(String itemModelType, String packName, Identifier itemModel, TestResultType result) {
+    private record TestResultUnit(String itemModelType, String packName, ResourceLocation itemModel, TestResultType result) {
         public String toPrint() {
             return "Pack " + packName + ": action " + String.join(" ", itemModelType, itemModel.toString(), result.toString());
         }