+++ src/main/java/com/danrus/rpf/mixin/load/BakingResultMixin.java
@@ -5,7 +5,7 @@ import com.danrus.rpf.duck.load.RpfBakingResult;
 import net.minecraft.client.renderer.item.ClientItem;
 import net.minecraft.client.renderer.item.ItemModel;
 import net.minecraft.client.resources.model.ModelBakery;
-import net.minecraft.resources.Identifier;
+import net.minecraft.resources.ResourceLocation;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 
@@ -16,30 +16,30 @@ import java.util.Map;
 public class BakingResultMixin implements RpfBakingResult {
 
     @Unique
-    private List<Map<Identifier, SignedItemModel>> modelsList;
+    private List<Map<ResourceLocation, SignedItemModel>> modelsList;
 
     @Unique
-    private List<Map<Identifier, ClientItem.Properties>> propertiesList;
+    private List<Map<ResourceLocation, ClientItem.Properties>> propertiesList;
 
     @Override
-    public ModelBakery.BakingResult rpf$setSignedItemModels(List<Map<Identifier, SignedItemModel>> models) {
+    public ModelBakery.BakingResult rpf$setSignedItemModels(List<Map<ResourceLocation, SignedItemModel>> models) {
         this.modelsList = models;
         return (ModelBakery.BakingResult) (Object) this;
     }
 
     @Override
-    public List<Map<Identifier, SignedItemModel>> rpf$getItemSignedModels() {
+    public List<Map<ResourceLocation, SignedItemModel>> rpf$getItemSignedModels() {
         return modelsList;
     }
 
     @Override
-    public RpfBakingResult rpf$setItemProperties(List<Map<Identifier, ClientItem.Properties>> properties) {
+    public RpfBakingResult rpf$setItemProperties(List<Map<ResourceLocation, ClientItem.Properties>> properties) {
         this.propertiesList = properties;
         return this;
     }
 
     @Override
-    public List<Map<Identifier, ClientItem.Properties>> rpf$getItemProperties() {
+    public List<Map<ResourceLocation, ClientItem.Properties>> rpf$getItemProperties() {
         return propertiesList;
     }
 }