+++ src/main/java/com/danrus/rpf/mixin/load/ModelManagerMixin.java
@@ -18,7 +18,7 @@ import net.minecraft.client.renderer.block.model.ItemModelGenerator;
 import net.minecraft.client.renderer.item.ClientItem;
 import net.minecraft.client.renderer.item.ItemModel;
 import net.minecraft.client.resources.model.*;
-import net.minecraft.resources.Identifier;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.util.profiling.Profiler;
 import net.minecraft.util.profiling.ProfilerFiller;
@@ -43,10 +43,10 @@ public abstract class ModelManagerMixin implements RpfModelManager {
 
     @Unique
 //    private List<Map<ResourceLocation, ItemModel>> rpf$bakedItemStackModels;
-    private List<Map<Identifier, SignedItemModel>> rpf$bakedItemStackSignetModels;
+    private List<Map<ResourceLocation, SignedItemModel>> rpf$bakedItemStackSignetModels;
 
     @Unique
-    private List<Map<Identifier, ClientItem.Properties>> rpf$itemProperties;
+    private List<Map<ResourceLocation, ClientItem.Properties>> rpf$itemProperties;
 
     @Unique
     private static final ClientItemInfoLoader.LoadedClientInfos EMPTY_LOADED_INFOS =
@@ -56,24 +56,24 @@ public abstract class ModelManagerMixin implements RpfModelManager {
     private ModelBakery.MissingModels missingModels;
 
     //? if <= 1.21.8 {
-    /*@Shadow
-    @Final
-    private AtlasSet atlases;
-    *///? } else {
     @Shadow
     @Final
+    private AtlasSet atlases;
+    //? } else {
+    /*@Shadow
+    @Final
     private AtlasManager atlasManager;
-    //?}
+    *///?}
 
     //? if <=1.21.8
-    //@Shadow private int maxMipmapLevels;
+    @Shadow private int maxMipmapLevels;
 
     @Shadow
     @Final
     private BlockColors blockColors;
 
     @Shadow
-    private static CompletableFuture<Map<Identifier, UnbakedModel>> loadBlockModels(ResourceManager resourceManager, Executor executor) {
+    private static CompletableFuture<Map<ResourceLocation, UnbakedModel>> loadBlockModels(ResourceManager resourceManager, Executor executor) {
         return null; // Shadowed implementation
     }
 
@@ -85,7 +85,7 @@ public abstract class ModelManagerMixin implements RpfModelManager {
     @Shadow
     protected abstract void apply(ModelManager.ReloadState reloadState
             //? if <=1.21.8
-            //, ProfilerFiller profiler
+            , ProfilerFiller profiler
     );
 
     @WrapOperation(
@@ -105,25 +105,25 @@ public abstract class ModelManagerMixin implements RpfModelManager {
             method = "net/minecraft/client/resources/model/ModelManager.method_65747(Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/lang/Void;)Lnet/minecraft/client/resources/model/ModelManager$ResolvedModels;",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/resources/model/ModelManager;discoverModelDependencies(Ljava/util/Map;Lnet/minecraft/client/resources/model/BlockStateModelLoader$LoadedModels;Lnet/minecraft/client/resources/model/ClientItemInfoLoader$LoadedClientInfos;)Lnet/minecraft/client/resources/model/ModelManager$ResolvedModels;")
     )
-    private static ModelManager.ResolvedModels rpf$wrapDiscovery(Map<Identifier, UnbakedModel> inputModels, BlockStateModelLoader.LoadedModels loadedModels, ClientItemInfoLoader.LoadedClientInfos loadedClientInfos, Operation<ModelManager.ResolvedModels> original) {
+    private static ModelManager.ResolvedModels rpf$wrapDiscovery(Map<ResourceLocation, UnbakedModel> inputModels, BlockStateModelLoader.LoadedModels loadedModels, ClientItemInfoLoader.LoadedClientInfos loadedClientInfos, Operation<ModelManager.ResolvedModels> original) {
         return rpf$discoverModelDependencies(inputModels, loadedModels, Rpf.rpf$currentItemLayersFuture.join());
     }
 
     @Unique
     private static final String redirectModelBakeryConstructorTarget =
             //? if 1.21.8
-            //"net/minecraft/client/resources/model/ModelManager.method_65753(Ljava/util/Map;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/Executor;Ljava/lang/Void;)Ljava/util/concurrent/CompletionStage;";
+            "net/minecraft/client/resources/model/ModelManager.method_65753(Ljava/util/Map;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/Executor;Ljava/lang/Void;)Ljava/util/concurrent/CompletionStage;";
             //? if 1.21.10
             //"net/minecraft/client/resources/model/ModelManager.method_65753(Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/Executor;Ljava/lang/Void;)Ljava/util/concurrent/CompletionStage;";
             //? if 1.21.11
-            "net/minecraft/client/resources/model/ModelManager.method_65753(Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/Executor;Ljava/lang/Void;)Ljava/util/concurrent/CompletionStage;";
+            //"net/minecraft/client/resources/model/ModelManager.method_65753(Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/CompletableFuture;Ljava/util/concurrent/Executor;Ljava/lang/Void;)Ljava/util/concurrent/CompletionStage;";
 
     @Unique
     private static final String redirectModelBakeryConstructorBakeryTarget =
             //? if 1.21.8
-            //"(Lnet/minecraft/client/model/geom/EntityModelSet;Ljava/util/Map;Ljava/util/Map;Ljava/util/Map;Lnet/minecraft/client/resources/model/ResolvedModel;)Lnet/minecraft/client/resources/model/ModelBakery;";
+            "(Lnet/minecraft/client/model/geom/EntityModelSet;Ljava/util/Map;Ljava/util/Map;Ljava/util/Map;Lnet/minecraft/client/resources/model/ResolvedModel;)Lnet/minecraft/client/resources/model/ModelBakery;";
             //? if >=1.21.10
-            "(Lnet/minecraft/client/model/geom/EntityModelSet;Lnet/minecraft/client/resources/model/MaterialSet;Lnet/minecraft/client/renderer/PlayerSkinRenderCache;Ljava/util/Map;Ljava/util/Map;Ljava/util/Map;Lnet/minecraft/client/resources/model/ResolvedModel;)Lnet/minecraft/client/resources/model/ModelBakery;";
+            //"(Lnet/minecraft/client/model/geom/EntityModelSet;Lnet/minecraft/client/resources/model/MaterialSet;Lnet/minecraft/client/renderer/PlayerSkinRenderCache;Ljava/util/Map;Ljava/util/Map;Ljava/util/Map;Lnet/minecraft/client/resources/model/ResolvedModel;)Lnet/minecraft/client/resources/model/ModelBakery;";
 
     @WrapOperation(
             method = redirectModelBakeryConstructorTarget,
@@ -131,11 +131,11 @@ public abstract class ModelManagerMixin implements RpfModelManager {
     )
     private static ModelBakery rpf$redirectModelBakeryConstructor
             //? if 1.21.8
-            //(EntityModelSet entityModelSet, Map unbakedBlockStateModels, Map clientInfos, Map resolvedModels, ResolvedModel missingModel, Operation<ModelBakery> original)
+            (EntityModelSet entityModelSet, Map unbakedBlockStateModels, Map clientInfos, Map resolvedModels, ResolvedModel missingModel, Operation<ModelBakery> original)
             //? if >=1.21.10
-            (EntityModelSet entityModelSet, MaterialSet materials, net.minecraft.client.renderer.PlayerSkinRenderCache playerSkinRenderCache, Map unbakedBlockStateModels, Map clientInfos, Map resolvedModels, ResolvedModel missingModel, Operation<ModelBakery> original)
+            //(EntityModelSet entityModelSet, MaterialSet materials, net.minecraft.client.renderer.PlayerSkinRenderCache playerSkinRenderCache, Map unbakedBlockStateModels, Map clientInfos, Map resolvedModels, ResolvedModel missingModel, Operation<ModelBakery> original)
     {
-        List<Map<Identifier, ClientItem>> rawLayers = new ArrayList<>();
+        List<Map<ResourceLocation, ClientItem>> rawLayers = new ArrayList<>();
         for (RpfClientItemInfoLoader.LoadedClientInfos layer : Rpf.rpf$currentItemLayersFuture.join()) {
             rawLayers.add(layer.contents());
         }
@@ -149,15 +149,15 @@ public abstract class ModelManagerMixin implements RpfModelManager {
         }
         return ((RpfModelBakery) original.call
                 //? if 1.21.8
-                //(entityModelSet, unbakedBlockStateModels, clientInfos, resolvedModels, missingModel)
+                (entityModelSet, unbakedBlockStateModels, clientInfos, resolvedModels, missingModel)
                 //? if >=1.21.10
-                (entityModelSet, materials, playerSkinRenderCache, unbakedBlockStateModels, clientInfos, resolvedModels, missingModel)
+                //(entityModelSet, materials, playerSkinRenderCache, unbakedBlockStateModels, clientInfos, resolvedModels, missingModel)
         ).rpf$setClientItems(rawLayers);
     }
 
     @Unique
     private static ModelManager.ResolvedModels rpf$discoverModelDependencies(
-            Map<Identifier, UnbakedModel> blockModels,
+            Map<ResourceLocation, UnbakedModel> blockModels,
             BlockStateModelLoader.LoadedModels loadedModels,
             List<RpfClientItemInfoLoader.LoadedClientInfos> itemLayers
     ) {
@@ -182,14 +182,14 @@ public abstract class ModelManagerMixin implements RpfModelManager {
     private void rpf$apply(
             ModelManager.ReloadState reloadState,
             //? if <= 1.21.8
-            //ProfilerFiller profiler,
+            ProfilerFiller profiler,
             CallbackInfo ci,
             @Local ModelBakery.BakingResult bakingResult
     ) {
         try {
             RpfBakingResult result = ((RpfBakingResult) (Object) bakingResult);
-            this.rpf$bakedItemStackSignetModels = result.rpf$getItemSignedModels().reversed(); // "reversed" to put vanilla RP down of list
-            this.rpf$itemProperties = result.rpf$getItemProperties().reversed();
+            this.rpf$bakedItemStackSignetModels = new ArrayList<>(result.rpf$getItemSignedModels().reversed());// "reversed" to put vanilla RP down of list
+            this.rpf$itemProperties = new ArrayList<>(result.rpf$getItemProperties().reversed());
             Rpf.getItemLogger().onReload();
         } catch (ClassCastException e) {
             throw new IllegalStateException("ModelBakery.BakingResult bakingResult is not instance of RpfBakingResult!");
@@ -200,12 +200,12 @@ public abstract class ModelManagerMixin implements RpfModelManager {
 
 
     @Override
-    public List<Map<Identifier, SignedItemModel>> rpf$getSignedModels() {
+    public List<Map<ResourceLocation, SignedItemModel>> rpf$getSignedModels() {
         return rpf$bakedItemStackSignetModels;
     }
 
     @Override
-    public List<Map<Identifier, ClientItem.Properties>> rpf$getItemPropertiesMaps() {
+    public List<Map<ResourceLocation, ClientItem.Properties>> rpf$getItemPropertiesMaps() {
         return this.rpf$itemProperties;
     }